backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
yahooquery>=2.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
exa-py>=1.0.0instagrapi>=2.0.0
pillow>=10.0.0
python-dateutil>=2.8.0